+++ nengo/tests/test_config.py
@@ -1,8 +1,7 @@
 import pytest
 
 import nengo
-import nengo.synapses
-from nengo.config import SupportDefaultsMixin
+from nengo.base import NengoObject
 from nengo.exceptions import ConfigError, ReadonlyError
 from nengo.params import Default, Parameter
 from nengo.utils.testing import ThreadedAssertion
@@ -140,7 +139,7 @@ def test_configstack():
         e2 = nengo.Ensemble(6, dimensions=1)
         excite = nengo.Connection(e1, e2)
         with nengo.Config(nengo.Connection) as inhib:
-            inhib[nengo.Connection].synapse = nengo.synapses.Lowpass(0.00848)
+            inhib[nengo.Connection].synapse = nengo.Lowpass(0.00848)
             inhibit = nengo.Connection(e1, e2)
     assert excite.synapse == nengo.Connection.synapse.default
     assert excite.transform.init == -1
@@ -216,10 +215,11 @@ def test_contains():
 
 
 def test_subclass_config():
-    class MyParent(SupportDefaultsMixin):
+    class MyParent(NengoObject):
         p = Parameter("p", default="baba")
 
         def __init__(self, p=Default):
+            super().__init__(None, None)
             self.p = p
 
     class MyChild(MyParent):
@@ -227,12 +227,12 @@ def test_subclass_config():
 
     with nengo.Config(MyParent) as cfg:
         cfg[MyParent].p = "value1"
-        a = MyChild()
+        a = MyChild(add_to_container=False)
         assert a.p == "value1"
 
     with nengo.Config(MyParent) as cfg:
         cfg[MyChild].p = "value2"
-        a = MyChild()
+        a = MyChild(add_to_container=False)
         assert a.p == "value2"
 
     # If any config entry in the current context fits with the object being
@@ -243,5 +243,5 @@ def test_subclass_config():
 
         with nengo.Config(MyParent) as cfg2:
             cfg2[MyParent].p = "value2"
-            a = MyChild()
+            a = MyChild(add_to_container=False)
             assert a.p == "value2"
